Apparently able to boot into FreeRTOS with TCP/IP enabled.
Honestly I'm not sure what I did to get this working. I took a weeklong diversion
to get the Debug Probe working (timer interrupts threw me for a loop) and when I
finally got that working, everything else worked too, Debug Probe or not. Don't
ask me. Anyway, it doesn't join a network yet, but it tries, and that effort counts
for something. :)